Oops, ForwardAuth would spin the CPU if it ever got disconnected.
This was sort of harmless, in that it just sucked a lot of CPU until the
next authentication request, but it was still silly.  Now we shut down the
forwarding connection cleanly in that case, so that the next auth request
can work.

We also send a keepalive after TCP_TIMEOUT_MS/2 milliseconds so the server
is less likely to disconnect us for no reason.