Add basic authcache support.
There are two levels of cache:

- accelerated: if a successful cache entry is < cache_accel_secs old, we
  don't even bother checking the server and just return success right away.
  This is great for speeding up things like apache_mod_pam.

- normal: we use the cache only if there's trouble connecting to the server.

The timeouts haven't been set to reasonable values and aren't configurable
yet.  The timeout for "accelerated" should probably be a few minutes (since
it's a bit of a security hole) while the timeout for "normal" isn't too bad,
and the comparable timeout in Windows XP seems to be infinite.

We don't persist the cache to disk yet.